requirements.txt
PyMySQL==1.1.1
python-dotenv==1.0.1
PyJWT==2.8.0
bcrypt==4.1.3
